Rewrote code

+++ views/captains.py
@@ -7,34 +7,35 @@ import settings
 from views.pick import PickView
 from views.menu import WinnerView
 
+from utils import BasicEmbed, CaptainEmbed
+
 
 class CaptainsView(discord.ui.View):
     def __init__(self,
+                 t_cap: discord.Member,
+                 e_cap: discord.Member,
                  players: list[discord.Member],
-                 captain_team: discord.Member,
-                 captain_enemy: discord.Member,
                  host: discord.Member,
                  game: discord.app_commands.Choice,
                  text: discord.TextChannel,
-                 voice: discord.VoiceChannel,
-                 emb: discord.Embed):
+                 voice: discord.VoiceChannel):
         super().__init__()
 
         self.players = players
-        self.captain_team = captain_team
-        self.captain_enemy = captain_enemy
+        self.t_cap = t_cap
+        self.e_cap = e_cap
         self.host = host
         self.game = game
 
         self.text = text
         self.voice = voice
-        self.embed = emb
         self.timeout = 60 * 60 * 5  # 5 hours
 
-        self.enemy_players = []
-        self.team_players = []
+        self.t_players = []
+        self.e_players = []
         self.vc_1 = None
         self.vc_2 = None
+        self.active_game = False
 
     @discord.ui.button(label='Начать', style=discord.ButtonStyle.green, custom_id='registration_btn')
     async def registration_callback(self, interaction: discord.Interaction, button: discord.ui.Button):
@@ -43,71 +44,42 @@ class CaptainsView(discord.ui.View):
         button.disabled = True
         await interaction.edit_original_response(view=self)
 
-        guild = interaction.guild
+        turn = False
+        data = {'caps': (self.t_cap, self.e_cap), 'teams': (self.t_players, self.e_players)}
+        n = len(self.players)
+        for i in range(n):
+            cap: discord.Member = data['caps'][turn]
+            team: list[discord.Member] = data['teams'][turn]
 
-        """TODO RECODE THIS MESS"""
-        for i in range(1):
-            # First Captain
-            view = PickView(self.players, self.captain_team)
-            await interaction.followup.send(f'{self.captain_team.mention}', view=view)
+            view = PickView(cap, self.players)
+            message: discord.WebhookMessage = await interaction.followup.send(f'{cap.mention}', view=view)
             await view.wait()
-            await self.text.purge(limit=1)
+            await message.delete()
 
-            picked_player = guild.get_member(int(view.picked_player))
-            self.team_players.append(picked_player)
-            self.players.remove(picked_player)
-
-            await interaction.edit_original_response(
-                embed=update(
-                    self.embed,
-                    self.captain_team,
-                    self.team_players,
-                    self.captain_enemy,
-                    self.enemy_players,
-                    self.players,
-                    self.host
-                ),
-                view=self
-            )
-
-            # Second Captain
-            view_2 = PickView(self.players, self.captain_enemy)
-            await interaction.followup.send(f'{self.captain_enemy.mention}', view=view_2)
-            await view_2.wait()
-            await self.text.purge(limit=1)
-
-            picked_player = guild.get_member(int(view_2.picked_player))
-            self.enemy_players.append(picked_player)
-            self.players.remove(picked_player)
+            player = interaction.guild.get_member(int(view.picked_player))
+            team.append(player)
+            self.players.remove(player)
 
             await interaction.edit_original_response(
-                embed=update(
-                    self.embed,
-                    self.captain_team,
-                    self.team_players,
-                    self.captain_enemy,
-                    self.enemy_players,
-                    self.players,
-                    self.host
-                ),
-                view=self
+                embed=CaptainEmbed(data['caps'], self.players, self.host, teams=data['teams']), view=self
             )
+            turn = not turn
 
         # Access to vc
-        self.team_players.append(self.captain_team)
-        self.enemy_players.append(self.captain_enemy)
+        self.t_players.append(self.t_cap)
+        self.e_players.append(self.e_cap)
 
-        self.vc_1 = await self.text.category.create_voice_channel(name=f'team_{self.captain_team}')
-        self.vc_2 = await self.text.category.create_voice_channel(name=f'team_{self.captain_enemy}')
+        self.vc_1 = await self.text.category.create_voice_channel(name=f'team_{self.t_cap}')
+        self.vc_2 = await self.text.category.create_voice_channel(name=f'team_{self.e_cap}')
         await self.vc_1.edit(sync_permissions=True)
         await self.vc_2.edit(sync_permissions=True)
 
-        for team, enemy in zip(self.team_players, self.enemy_players):
+        for team, enemy in zip(self.t_players, self.e_players):
             await self.vc_1.set_permissions(target=enemy, overwrite=discord.PermissionOverwrite(connect=False))
             await self.vc_2.set_permissions(target=team, overwrite=discord.PermissionOverwrite(connect=False))
 
         # Move players to their team voice
-        for team, enemy in zip(self.team_players, self.enemy_players):
+        for team, enemy in zip(self.t_players, self.e_players):
             await team.move_to(self.vc_1)
             await enemy.move_to(self.vc_2)
 
@@ -115,20 +87,22 @@ class CaptainsView(discord.ui.View):
             alphabet = string.ascii_letters + string.digits
             lobby = f'discord.gg/5x5_{random.randint(0, 100)}'
             password = ''.join(secrets.choice(alphabet) for _ in range(6))
-            embed = discord.Embed(
-                title='Лобби в Dota 2', description=f'Название: **{lobby}**\nПароль: **{password}**', colour=2829617
-            )
-            await interaction.followup.send(embed=embed)
+            await interaction.followup.send(embed=BasicEmbed(lobby, password))
 
+        self.active_game = True
         await interaction.followup.send(f"Удачной игры!")
 
     @discord.ui.button(label='Завершить', style=discord.ButtonStyle.red, custom_id='exit_btn')
     async def exit_callback(self, interaction: discord.Interaction, button: discord.ui.Button):
-        winner = WinnerView(
-            self.captain_team, self.captain_enemy, self.team_players, self.enemy_players, self.game, self.host
-        )
-        await interaction.response.send_message(view=winner, ephemeral=True)
-        await winner.wait()
+        if not self.active_game:
+            await interaction.response.send_message(
+                embed=BasicEmbed('Не активная игра', 'Дождитесь окончания пика для завершения'), ephemeral=True
+            )
+            return
+
+        view = WinnerView(self.t_cap, self.e_cap, self.t_players, self.e_players, self.game, self.host)
+        await interaction.response.send_message(view=view, ephemeral=True)
+        await view.wait()
 
         await interaction.channel.send('Завершено')
         await self.vc_1.delete()
@@ -140,20 +114,3 @@ class CaptainsView(discord.ui.View):
         if interaction.user.guild_permissions.administrator or settings.CLOSER_ROLE in interaction.user.roles:
             return True
         return False
-
-
-def update(embed: discord.Embed, c_team, team_player, c_enemy, enemy_player, remaining_player, host):
-    embed.clear_fields()
-
-    msg = '\n'.join(player.mention for player in remaining_player)
-    team = '\n'.join(player.mention for player in team_player)
-    enemy = '\n'.join(player.mention for player in enemy_player)
-
-    if msg != '':
-        embed.add_field(name='Игроки', value=msg)
-
-    embed.add_field(name=f'team_{c_team}', value=team)
-    embed.add_field(name=f'team_{c_enemy}', value=enemy)
-    embed.set_footer(text=f'Hosted by {host.name}')
-
-    return embed
